requirements.txt
Django==1.9.2
Pygments==2.1
blessings==1.6
bpython==0.14.1
braintree==3.24.0
cssselect==0.9.1
curtsies==0.1.23
django-polymorphic==0.8.1
django-registration-redux==1.3
djangorestframework==3.3.2
greenlet==0.4.9
lxml==3.4.2
mistune==0.5.1
mock==1.0.1
model-mommy==1.2.6
psycopg2==2.6
py==1.4.31
pytest==2.8.7
pytest-django==2.9.1
requests==2.9.1
six==1.10.0
